+++ subject/chapter01/exam.go
@@ -1,4 +1,4 @@
-package exam01
+package chapter01
 
 import "github.com/kurupeku/hello-golang/helper"
